+++ src/constants.ts
@@ -2,7 +2,11 @@ export const EXTENSION_NAME = 'sftp';
 
 export const CONFIG = 'sftp.config.default';
 
-export const WORKSPACE_TRIE_TOKEN = '@workroot';
+export const VSCODE_FOLDER = '.vscode';
+export const DEPRECATED_CONGIF_FILENAME = '.sftpConfig.json';
+export const CONGIF_FILENAME = 'sftp.json';
+
+export const CONFIG_GLOB_PATTERN = `**/${VSCODE_FOLDER}/{${CONGIF_FILENAME},${DEPRECATED_CONGIF_FILENAME}}`;
 
 export const SYNC_TO_REMOTE = 'sftp.sync.remote';
 export const SYNC_TO_LOCAL = 'sftp.sync.local';